+++ MySecInfo.h
@@ -53,4 +53,4 @@ private:
 };
 
 _Check_return_ wstring GetTextualSid(_In_ PSID pSid);
-_Check_return_ HRESULT SDToString(_In_count_(cbBuf) LPBYTE lpBuf, ULONG cbBuf, eAceType acetype, _In_ wstring& SDString, _In_ wstring& sdInfo);
+_Check_return_ HRESULT SDToString(_In_count_(cbBuf) LPBYTE lpBuf, size_t cbBuf, eAceType acetype, _In_ wstring& SDString, _In_ wstring& sdInfo);
